chore: use base bottom sheet in expansive modal
